Home

TensorMixedStates (TMS) is a Julia library to make simulations of closed or open quantum systems using Matrix Product States representations.

Features

TMS uses Matrix Product State representations for the density matrix of the system. It proposes a large set of features: manipulations of systems and states, a rich set of sites and operators easily extensible by the user, powerful algorithms: computation of ground states using DMRG, Hamiltonian and Lindbladian evolution with TDVP (and others), applications of gates (including noisy gates).

Being based on ITensor, TMS delivers high performance computations and naturally runs in parallel.

The interface is user friendly. In particular, it features a very expressive syntax for operators allowing easy definitions of operators such as observables, gates, Hamiltonians or Lindbladians. Moreover the optional high level interface allows the writing of simple simulations in a few lines of code.

References

TMS is described in the following article, published in SciPost Physics Codebases. Please cite both the article and the codebase release it documents, which is the convention SciPost asks for:

Jérôme Houdayer and Grégoire Misguich, TensorMixedStates: A Julia library for simulating pure and mixed quantum states using matrix product states, SciPost Phys. Codebases 72 (2026).

Jérôme Houdayer and Grégoire Misguich, Codebase release 1.28 for TensorMixedStates, SciPost Phys. Codebases 72-r1.28 (2026).

In BibTeX:

@article{TensorMixedStates,
  title     = {{TensorMixedStates}: A {Julia} library for simulating pure and mixed
               quantum states using matrix product states},
  author    = {Houdayer, J{\'e}r{\^o}me and Misguich, Gr{\'e}goire},
  journal   = {SciPost Phys. Codebases},
  pages     = {72},
  year      = {2026},
  publisher = {SciPost},
  doi       = {10.21468/SciPostPhysCodeb.72},
}

@article{TensorMixedStatesRelease,
  title     = {{Codebase} release 1.28 for {TensorMixedStates}},
  author    = {Houdayer, J{\'e}r{\^o}me and Misguich, Gr{\'e}goire},
  journal   = {SciPost Phys. Codebases},
  pages     = {72-r1.28},
  year      = {2026},
  publisher = {SciPost},
  doi       = {10.21468/SciPostPhysCodeb.72-r1.28},
}

Acknowledgements and feedback

If TMS helped produce the data of a publication, a word about it in your acknowledgements would be warmly appreciated. It costs you a single line, and it is what makes it possible to keep the library developed and maintained.

We would also be delighted to hear from you directly, and you should not hesitate for a moment. What are you doing with TMS? What works well, what gets in your way, what do you need that is not there yet? The contact address is the one given in the article. Such messages are read with care, and they are what shapes what comes next.

Installation

To use TMS, you need to have Julia installed on your system. Installing julia is usually easy and fast, see The Julia Programming Language for instructions. TMS requires at least Julia version 1.10.5 to run.

To install TMS in Julia, launch the Julia interface (by typing 'julia' on the command line) and type

]add TensorMixedStates

Note the "]" required to enter julia package management system.

After downloading TMS, Julia will automatically compile and install it. This process usually takes a couple of minutes and does not require interactions on the user part.

BLAS backend

Most of the running time of TMS is spent in the tensor contractions of ITensors, which themselves call BLAS. Julia ships with OpenBLAS and TMS uses it as it comes: switching the BLAS backend affects the whole Julia session, so that choice is left to you rather than made by a library you load.

On x86_64 machines running Linux or Windows, Intel's MKL is often noticeably faster on these contractions. To use it, add MKL to your project and load it before TMS:

using MKL
using TensorMixedStates

MKL is not distributed for macOS nor for ARM machines, where OpenBLAS is the only option.

Using TMS

To use TMS in your code you need to write

using TensorMixedStates

Julia script names are usually written with a .jl extension. Once you have written your script, you can execute it with

julia my_script.jl

to run it on a single processor, or

julia --threads=4 my_script.jl
julia --threads=auto my_script.jl

to use multi-threading (see julia documentation for more details on multi-threading).

You can also use TMS in the interactive julia interpreter.

Bug reports and feature requests

TMS is still in development and certainly contains bugs. If you think you have found one please report it on the Github page.

Feature requests may also be sent on the Github page.

Documentation

You are currently reading it!

You can have access to inline documentation on TMS at the julia prompt simply by typing "?" followed by the function name or type name you are interested in. For example

?runTMS

For this to work you must have first imported TMS with

using TensorMixedStates

Learning about matrix product states

For an introduction to matrix product states and tensor networks, see the lecture notes of the course Grégoire Misguich gave at the 9th Les Houches summer school on Computational Physics: Open Quantum Systems, in June 2026, available on arXiv:2606.24803. Their repository holds the sixteen Julia examples the notes use, some of them written with TMS.

Examples

Working examples are presented in the folder examples in the repository.

Module

TensorMixedStates — Module

A module to make numerical simulations of closed or open quantum systems using Matrix Product States

Index

TensorMixedStates

TensorMixedStates.Bosons

TensorMixedStates.Electrons

TensorMixedStates.Fermions

TensorMixedStates.Qbosons

TensorMixedStates.Qubits

TensorMixedStates.Qudits

TensorMixedStates.Spins

TensorMixedStates.Tjs

TensorMixedStates.F

TensorMixedStates.Hermiticity

TensorMixedStates.HermiticityError

TensorMixedStates.Id

TensorMixedStates.Linkdim

TensorMixedStates.MaxLinkdim

TensorMixedStates.MemoryUsage

TensorMixedStates.Norm

TensorMixedStates.Purity

TensorMixedStates.Renyi2

TensorMixedStates.Swap

TensorMixedStates.Trace

TensorMixedStates.Trace2

TensorMixedStates.TraceError

TensorMixedStates.fermionic_op

TensorMixedStates.involution_op

TensorMixedStates.plain_op

TensorMixedStates.selfadjoint_op

ITensors.Index

TensorMixedStates.AbstractSite

TensorMixedStates.Algo

TensorMixedStates.ApproxW

TensorMixedStates.ApproxWObserver

TensorMixedStates.Boson

TensorMixedStates.Check

TensorMixedStates.CreateState

TensorMixedStates.Data

TensorMixedStates.Dissipator

TensorMixedStates.Dmrg

TensorMixedStates.DmrgObserver

TensorMixedStates.Electron

TensorMixedStates.Evolve

TensorMixedStates.Evolver

TensorMixedStates.Fermion

TensorMixedStates.Gate

TensorMixedStates.Gates

TensorMixedStates.GenericOp

TensorMixedStates.GroundState

TensorMixedStates.IndexedOp

TensorMixedStates.Left

TensorMixedStates.Limits

TensorMixedStates.LoadState

TensorMixedStates.Measure

TensorMixedStates.Mixed

TensorMixedStates.Op

TensorMixedStates.OpType

TensorMixedStates.Operator

TensorMixedStates.PartialTrace

TensorMixedStates.Phases

TensorMixedStates.PreMPO

TensorMixedStates.Proj

TensorMixedStates.Pure

TensorMixedStates.Qboson

TensorMixedStates.Qubit

TensorMixedStates.Qudit

TensorMixedStates.RandomState

TensorMixedStates.Right

TensorMixedStates.SaveState

TensorMixedStates.SetState

TensorMixedStates.SimData

TensorMixedStates.SimpleOp

TensorMixedStates.Simulation

TensorMixedStates.Spin

TensorMixedStates.State

TensorMixedStates.State

TensorMixedStates.StateFunc

TensorMixedStates.SteadyState

TensorMixedStates.SysIndex

TensorMixedStates.System

TensorMixedStates.Tdvp

TensorMixedStates.TdvpObserver

TensorMixedStates.TimeFunc

TensorMixedStates.Tj

TensorMixedStates.ToMixed

TensorMixedStates.Weaken

Base.length

Base.length

Base.mod

Base.truncate

ITensorMPS.dmrg

ITensorMPS.expect

ITensorMPS.maxlinkdim

ITensorMPS.sample

ITensorMPS.tdvp

ITensors.SiteTypes.state

ITensors.apply

ITensors.dag

ITensors.dag

ITensors.flux

ITensors.flux

ITensors.inner

LinearAlgebra.norm

LinearAlgebra.normalize

NDTensors.dim

NDTensors.matrix

TensorMixedStates.:⊗

TensorMixedStates.:⊗

TensorMixedStates.DataToFrame

TensorMixedStates.EE

TensorMixedStates.EntanglementEntropy

TensorMixedStates.Fidelity

TensorMixedStates.MutualInfoRenyi2

TensorMixedStates.Mutual_Info_Renyi2

TensorMixedStates.Overlap

TensorMixedStates.Phase

TensorMixedStates.SubRenyi2

TensorMixedStates.Sumd

TensorMixedStates.Variance

TensorMixedStates.approx_W

TensorMixedStates.circle_graph

TensorMixedStates.complete_graph

TensorMixedStates.conserve_string

TensorMixedStates.controlled

TensorMixedStates.create_graph_state

TensorMixedStates.data_to_frame

TensorMixedStates.entanglement_by_sector

TensorMixedStates.entanglement_entropy

TensorMixedStates.expect1

TensorMixedStates.expect2

TensorMixedStates.fidelity

TensorMixedStates.get_sim_file

TensorMixedStates.graph_base_size

TensorMixedStates.graph_state

TensorMixedStates.has_fermionic

TensorMixedStates.hermitianize

TensorMixedStates.hermiticity

TensorMixedStates.hs_fidelity

TensorMixedStates.identity_operator

TensorMixedStates.isfermionic

TensorMixedStates.line_graph

TensorMixedStates.load_state

TensorMixedStates.log_msg

TensorMixedStates.make_approx_W1

TensorMixedStates.make_approx_W2

TensorMixedStates.make_mpo

TensorMixedStates.measure

TensorMixedStates.mix

TensorMixedStates.mutual_info_renyi2

TensorMixedStates.named

TensorMixedStates.output

TensorMixedStates.parity

TensorMixedStates.partial_trace

TensorMixedStates.renyi2

TensorMixedStates.runTMS

TensorMixedStates.save_state

TensorMixedStates.simplify

TensorMixedStates.square_lattice

TensorMixedStates.steady_state

TensorMixedStates.string_state

TensorMixedStates.strong

TensorMixedStates.symmetries

TensorMixedStates.tensor

TensorMixedStates.trace

TensorMixedStates.trace2

TensorMixedStates.variance

TensorMixedStates.weaken

TensorMixedStates.@create_site_module

TensorMixedStates.@def_operators

TensorMixedStates.@def_states