Measurements

measure

TensorMixedStates.measure — Function

measure(state, args[, t])
measure(state, measure[, t])
measure(state, [measures...][, t])

compute the requested measurements on the given state and simulation time

It is more efficient to ask all required measurements in one call

Examples

measure(state, X(1))     # compute observable X(1)
measure(state, X)        # compute observable X on all sites
measure(state, (X, Y))   # compute correlations XY on all pairs of sites
measure(state, Check("check", X(1)X(2), t->sin(2t)), 0.8) # compute and check the given observable against a computed value
measure(state, [X(2), Y, (X, Y)]) # several measures together

TensorMixedStates.StateFunc — Type

struct StateFunc
StateFunc(name, func)

a data type to represent a function of State. This is used by measure. This how Trace, Purity ... are implemented.

Examples

Trace = StateFunc("Trace", trace)
measure(state, Trace)
measures = "data.dat" => Trace

TensorMixedStates.Check — Type

struct Check
Check(name, obs1, obs2[, tol])

a measurement that checks the equality between two measurements. It throws an error if the difference is larger than tol.

TensorMixedStates.Measure — Type

struct Measure
Measure(args...)

a data type to hold a set of measurements, used internally

ITensorMPS.expect — Function

expect(::State, obs)

Compute expectation values of obs on the given state.

Examples

expect(state, X(1)*Y(2) + Y(1)*Z(3))
expect(state, [X(1)*Y(2), X(3), Z(1)*X(2)])

TensorMixedStates.expect1 — Function

expect1(::State, op)

Compute the expectation values of the given operators on all sites.

Examples

expect1(state, X)
expect1(state, [X, Y, Z])

TensorMixedStates.expect2 — Function

expect2(::State, op_pairs)

Compute the 2-point correlations of the given pairs of operators on all sites.

Examples

expect2(state, (X, X))
expect2(state, [(X, Y), (X, Z), (Y, Z)])
...

TensorMixedStates.entanglement_entropy — Function

entanglement_entropy(::State, ::Int)

Return the entanglement entropy of the given state at the given site. Also return the associated spectrum.

TensorMixedStates.renyi2 — Function

renyi2(::State)
renyi2(::State, ::Vector{Int})

renyi2 returns the Renyi entropy of order 2 of the state. This is 0. for pure representations. When given an array of positions give the Renyi entropy of corresponding substate.

TensorMixedStates.mutual_info_renyi2 — Function

mutual_info_renyi2(state::State, cut::Int)
mutual_info_renyi2(state::State, a::Vector{Int})

return an approximation of the mutual information using renyi2 entropy. You define the two parts either by giving the position of the cut between the left and right parts or by giving the list of positions for one of the parts.

TensorMixedStates.Trace — Constant

Trace

a state function to measure the trace of the system (density matrix). See also StateFunc and trace.

TensorMixedStates.TraceError — Constant

TraceError

a state function to measure the deviation to trace 1 of the system (density matrix). See also StateFunc and Trace. This is good way to measure the coherence of a simulation as numerical inaccuracies tend to change the trace of the density matrix (which should stay 1)

TensorMixedStates.Trace2 — Constant

Trace2
Purity

state functions to measure the trace of the square of the density matrix. See also StateFunc and trace2.

TensorMixedStates.Purity — Constant

Trace2
Purity

state functions to measure the trace of the square of the density matrix. See also StateFunc and trace2.

TensorMixedStates.Hermitianity — Constant

Hermitianity

a state function to measure the degree of hermitianity of the density matrix. Return 1 if density matrix is Hermitian, 0 for anti Hermitian, in beyween otherwise. Sea also StateFunc and hermitianity

TensorMixedStates.HermitianityError — Constant

HermitianityError

a state function to measure the deviation of the Hermitianity from 1. See StateFunc, Hermitianity and hermitianity.

TensorMixedStates.Renyi2 — Constant

Renyi2

a state function to measure the Renyi-2 entropy of the system. See also StateFunc and renyi2.

TensorMixedStates.SubRenyi2 — Function

SubRenyi2([positions...])

a state function to measure the Renyi-2 entropy of a subsystem describe by the positions given. See also StateFunc, Renyi2 and renyi2.

TensorMixedStates.Mutual_Info_Renyi2 — Function

Mutual_Info_Renyi2(link)
Mutual_Info_Renyi2([positions...])

a state function to measure the Renyi-2 mutual information of the given subsystems. See also StateFunc and mutual_info_renyi2.

TensorMixedStates.EE — Function

EE(pos)
EE(pos, spectrum)

a state function to measure entanglement entropy / OSEE and associated spectrum. See also StateFunc and entanglement_entropy.

TensorMixedStates.Linkdim — Constant

Linkdim

a state function to measure the maximum bond dimension. See also StateFunc and maxlinkdim.

TensorMixedStates.MemoryUsage — Constant

MemoryUsage

a state function to measure the memory used by the state. See also StateFunc.

Types of measurements

There are five types of measurements we can ask to measure

With an indexed operator, we get the expectation of the corresponding observable

mesure(state, X(1)X(2)Z(4))

With a generic operator (acting on one site only), we get the expectation of the operator on each site

measure(state, X)

With a couple of generic operators, we get the correlation matrix of those observators

measure(state, (X, Y))

There are a some state functions predefined:

Trace              # returns the trace of the state
Trace2                  # returns the trace of the square of the state (alternate name: Purity)
TraceError              # returns 1 - trace, usefull for monitoring trace deviations
Renyi2                  # returns the Renyi entropy of order 2 of the state
SubRenyi2(sub)          # returns the Renyi entropy of order 2 of the subsystem (sub is a vector containing the indices of the sites of the subsystem)
Mutual_Info_Renyi2(sub) # returns the Renyi2 mutual information of the two subsystem (you either give one subsystem as a vector of indices or a splitting link)
EE(l, n)                # returns entanglement entropy at site l and first n singular values
Hermitianity            # returns 1 if density matrix is really Hermitian and down to 0 for anti Hermitian density matrix
HermitianityError       # return 1 - Hermitianity for monitoring hermitianity deviation

They are used like this

measure(state, TraceError)

New state functions may be defined by

stfunc = StateFunc(name, state->... )

Checks can be performed (useful for coherence tests)

measure(state, Check(name, obs1, obs2))      # returns 3 values ob1, obs2 and |obs2 - obs1|
messure(state, Check(name, obs1, obs2, tol)) # if |obs2 - obs1|>tol throw an error
measure(state, Check(name, obs1, obs2), t)

In Check, obs may also be constants, vectors and function of time (like t -> sin(t)), in this case the simulation time must be fed to measure as 3rd argument.

Output

TensorMixedStates.output — Function

output(::Simulation, [ filename => measure1, ... ])

compute the given measurements on a simultation and output them to the associated file or dict

filenames are interpreted by get_sim_file (see there for special values)

Examples

output(sim, "file" => [X, X(1)Y(2), (X, Y)])
output(sim, [ "file1" => [X, Y(2)], "file2" => Trace])

TensorMixedStates.log_msg — Function

log_msg(::Simulation, text)

log the given message on the "log" file of the simulation