Operators

Usage

There are two kinds of operators: generic (like X) and indexed (like X(3)). Indexed operators are applied to specific site numbers.

Operators can be used to defined Hamiltonians, for example

hamiltonian = - j * sum(X(i)X(i+1) + Y(i)Y(i+1) for i in 1:n-1) - h * sum(Z(i) for i in 1:n)

or Lindbladian dissipators like

dissipators = sum(Dissipator(Sp)(i) for i in 1:n)

to build Lindbladian

lindbladian = -im * hamiltonian + dissipators

Note the factor -im for the Hamiltonian.

Operators can be used to define quantum gates like

gates = H(1)Swap(1, 2)H(1)

Noisy gates can be defined using the Gate constructor, for example

noisygate = 0.7Gate(Id) + 0.1Gate(X) + 0.1Gate(Y) + 0.1Gate(Z)

Operators can be used to define observables

obs = X(1)X(2)Z(3)

Reference

Complex operators can be build from a rich set of functions, for example

Rxy(t) = exp(-im * t * (X⊗X + Y⊗Y) / 4)

Operators can be added and multiplied using usual operators (+, -, *, /, ^).

TensorMixedStates.Pure — Type

type Pure <: PM
Pure()

correspond to pure quantum representation

TensorMixedStates.Mixed — Type

type Mixed <: PM
Mixed()

correspond to mixed quantum representation

TensorMixedStates.Op — Type

Op{R <: PM, T <: GI, N}

the type of all operators.

Type parameters

R: is Pure or Mixed (the type of the representations on which the operator may be applied)

T: ig Generic or Indexed

N: the number of sites on which the operator must be applied (set to 1 for indexed operators)

TensorMixedStates.Operator — Type

type Operator{N} <: GenericOp{Pure, N}

the type of base operators (like X, Swap, C ...), N is the number of sites on which it may be applied.

Example

Operator{1}("X", Nothing, involution_op)     a base operator whose value is predefined by the sites
Operator{1}("Z", [1 0 ; 0 -1], involution_op)
Operator{2}("Swap", [ 1 0 0 0 ; 0 0 1 0 ; 0 1 0 0 ; 0 0 0 1], involution_op)
Operator{1}("Sx", (Sp + Sm) / 2, selfadjoint_op)
Operator{1}("C", [0 1 ; 0 0], fermionic_op)

TensorMixedStates.AtIndex — Type

struct AtIndex{R, N} <: IndexedOp{R}

represent an indexed operator (like X(1) or Swap(2, 4))

TensorMixedStates.:⊗ — Function

op1 ⊗ op2

tensor product for generic operators, alternative syntax: tensor(op1, op2)

Examples

controlled(op) = Proj("Up") ⊗ Id + Proj("Dn") ⊗ op
Rxy(t) = exp(im * t * (X ⊗ X + Y ⊗ Y) / 4)

::System ⊗ ::System
tensor(::System, ::System)

create the tensorial product of two systems

TensorMixedStates.Proj — Type

Proj(state)

an operator to project on the given state

Examples

Proj("Up")
Proj([1, 0])

TensorMixedStates.Dissipator — Type

Dissipator(op)

a Lindbladian dissipator based on op to be used in evolver for time evolution

Examples

Dissipator(Sp)
Dissipator(0.1 * C)
Dissipator(Z ⊗ Z)
Dissipator(0.9X⊗Sm⊗X + 0.1Y⊗Sm⊗Y)

TensorMixedStates.Gate — Type

Gate(op)

a generic operator acting as a gate on states in mixed representation. Usefull for building noisy gates

Examples

G = 0.9 * Gate(Id) + 0.1 Gate(X)

ITensors.dag — Method

dag(::GenericOp)

adjoint for generic operators

TensorMixedStates.isfermionic — Function

isfermionic(::GenericOp)

return whether a generic operator is fermionic

NDTensors.matrix — Function

matrix(a::GenericOp, site::AbstractSite...)

return the matrix of a generic operator for the given sites. If sites are all identical, you may give only one

Examples

matrix(X, Qubit())
matrix(Swap, Qubit())
matrix(X⊗A, Qubit(), Boson(2))

TensorMixedStates.tensor — Function

tensor(a::GenericOp, site::AbstractSite...)

return the ITensor of a generic operator for the given sites. If sites are all identical, you may give only one

Examples

tensor(X, Qubit())
tensor(Swap, Qubit())
tensor(X⊗A, Qubit(), Boson(2))

tensor(::System, ::AtIndex)

returns a tensor representing the given simple indexed operator acting on this system

TensorMixedStates.simplify — Function

simplify(op::Op; expand=false)

simplifies an operator, this is used internally by functions creating MPOs. expand tells wheteher to expand explicit operators with their definition