Operators

Usage

There are two kinds of operators: generic (like X) and indexed (like X(3)). Indexed operators are applied to specific site numbers.

Operators can be used to defined Hamiltonians, for example

hamiltonian = - j * sum(X(i)X(i+1) + Y(i)Y(i+1) for i in 1:n-1) - h * sum(Z(i) for i in 1:n)

or Lindbladian dissipators like

dissipators = sum(Dissipator(Sp)(i) for i in 1:n)

to build Lindbladian

lindbladian = -im * hamiltonian + dissipators

Note the factor -im for the Hamiltonian.

Operators can be used to define quantum gates like

gates = H(1)Swap(1, 2)H(1)

Noisy gates can be defined using the Gate constructor, for example

noisygate = 0.7Gate(Id) + 0.1Gate(X) + 0.1Gate(Y) + 0.1Gate(Z)

Operators can be used to define observables

obs = X(1)X(2)Z(3)

Reference

Complex operators can be build from a rich set of functions, for example

Rxy(t) = exp(-im * t * (X⊗X + Y⊗Y) / 4)

Operators can be added and multiplied using usual operators (+, -, *, /, ^).

TensorMixedStates.Pure — Type

type Pure
Pure()

correspond to pure quantum representation

TensorMixedStates.Mixed — Type

type Mixed
Mixed()

correspond to mixed quantum representation

TensorMixedStates.ExprOp — Type

abstract type ExprOp{T, N}

the abstract type of operators, T is Pure or Mixed and N is an Int for generic operators (the number of sites on which the operator may apply) and IndexOp for indexed operators

SimpleOp = ExprOp{Pure, 1}
ExprIndexed{T} = ExprOp{T, IndexOp}

TensorMixedStates.Operator — Type

type Operator{T, N} <: ExprOp{T, N}

the type of base operators (like X, Swap, C ...), T is Pure or Mixed and N is an Int.

Example

Operator("X")                      a base operator whose value is predefined by the sites
Operator("Z", [1 0 ; 0 -1])
Operator{Pure, 2}("Swap", [ 1 0 0 0 ; 0 0 1 0 ; 0 1 0 0 ; 0 0 0 1])
Operator("CX", controlled(X))
Operator("Sx", (Sp + Sm) / 2)
Operator("C", [0 1 ; 0 0], true)   a fermionic operator

TensorMixedStates.Indexed — Type

type Indexed{T, N} <: ExprIndexed{T}

represent a base indexed operator (like X(1) or Swap(2, 4))

TensorMixedStates.:⊗ — Function

op1 ⊗ op2

tensor product for generic operators, alternative syntax: tensor(op1, op2)

Examples

controlled(op) = Proj("Up") ⊗ Id + Proj("Dn") ⊗ op
Rxy(t) = exp(im * t * (X ⊗ X + Y ⊗ Y) / 4)

::System ⊗ ::System
tensor(::System, ::System)

create the tensorial product of two systems

TensorMixedStates.Proj — Type

Proj(state)

an operator to project on the given state

Examples

Proj("Up")
Proj([1, 0])

TensorMixedStates.Dissipator — Type

Dissipator(op)

a Lindbladian dissipator based on op to be used in evolver for time evolution

TensorMixedStates.Gate — Type

Gate(op)

a generic operator acting as a gate on states in mixed representation. Usefull for building noisy gates

Examples

G = 0.9 * Gate(Id) + 0.1 Gate(X)

ITensors.dag — Method

dag(::ExprOp)

adjoint for generic operators

TensorMixedStates.fermionic — Function

fermionic(a::ExprOp)

return whether an operator is fermionic or not

NDTensors.matrix — Function

matrix(a::ExprOp, site::AbstractSite...)

return the matrix of a generic operator for the given sites. If sites are all identical, you may give only one

Examples

matrix(X, Qubit())
matrix(Swap, Qubit())
matrix(X⊗A, Qubit(), Boson(2))

TensorMixedStates.tensor — Function

tensor(a::ExprOp, site::AbstractSite...)

return the ITensor of a generic operator for the given sites. If sites are all identical, you may give only one

Examples

tensor(X, Qubit())
tensor(Swap, Qubit())
tensor(X⊗A, Qubit(), Boson(2))

tensor(::System, ::Indexed)

returns a tensor representing the given base indexed operator acting on this system

TensorMixedStates.simplify — Function

simplify(op)

simplify indexed operator

TensorMixedStates.insertFfactors — Function

insertFfactors

insert the Jordan-Wigner F operators where needed

TensorMixedStates.process — Function

process(op)

simplify indexed operator and insert Jordan-Wigner F factors where needed