Systems and States

Systems

TensorMixedStates.System — Type

type System

represent a quantum system

Fields

sites::Vector{<:AbstractSite}: sites of the system

pure_indices::Vector{Index}: Indices for pure representations

mixed_indices::Vector{Index}: Indices for mixed representations

Examples

System(10, Qubit())
System([Qubit(), SpinOne(), Qubit(), Boson(5)])

Indexation

system[i]          gives site i
system[Pure(), i]  gives pure index i
system[Mixed(), i] gives mixed index i

Base.length — Method

length(::System)

return the number of sites in the system

States

TensorMixedStates.Limits — Type

A type to hold MPS limits

Fields

cutoff: the cutoff under which singular values are neglected

maxdim: the maximum bond dimension

TensorMixedStates.State — Type

struct State
State(Pure()|Mixed(), system, states)
State(::State, ::MPS)

represent the complete state of the simulated quantum system

Fields

type::Union{Pure, Mixed}: pure or mixed representation

system::System: system description

state::MPS: system state

preobs::PreObs: preprocessing data for computing observables

Examples

State(Pure(), system, "Up")
State(Mixed(), system, ["Up", "Dn", "Up"])
State(Mixed(), system, "FullyMixed")
State(Pure(), system, [1, 0])
State(state, mps)        returns a new state with the same system but a new mps

Operations

states can be added, substracted and multiplied by numbers

Base.length — Method

length(::State)

return the number of sites in the state

ITensorMPS.maxlinkdim — Method

maxlinkdim(::State)

return the maximum link dimension in the state

TensorMixedStates.mix — Method

mix(::State)

transform a pure representation into a mixed representation

Base.truncate — Method

truncate(::State; limits::Limits)

apply the truncation to the given state

TensorMixedStates.trace — Method

trace(::State)

Return the trace of the system, mostly usefull for mixed representations. This should be one.

TensorMixedStates.trace2 — Method

trace2(::State)

Return the trace of the square density matrix, mostly usefull for mixed representations. This is one for pure representations.

LinearAlgebra.norm — Method

norm(::State)

Return the norm of the state, mostly usefull for pure representations. This should be one for pure representation.

ITensors.dag — Method

dag(::State)

adjoint of density matrix for mixed representation

TensorMixedStates.symmetrize — Method

symmetrize(::State)

symmetrize state so that it is hermitian (only for mixed state)

TensorMixedStates.normsym — Function

normsym(::State)

equivalent to normalyze(symmetrize(state))

TensorMixedStates.random_state — Function

random_state(Pure()|Mixed(), ::System, linkdims::Int)
random_state(Pure()|Mixed(), ::State, linkdims::Int)

Return a random state, with the specified link dimension. If a State is given, randomize the given state.

Simulations

TensorMixedStates.Simulation — Type

Simulation(state[, time = 0])
Simulation(sim, state[, time = sim.time])

A type to represent simulation data ans store time and file data. It is used and returned by runTMS. The first form creates a simulation object. The second updates the state in the simulation object. (see also get_sim_file)

Most functions applicable to States can be applied to Simulations

Fields

state : the state of the system

time : the simulation time

output : if not nothing an io where to redirect output

files : a dictionary holding io where to write data

formats : format info for the output

TensorMixedStates.get_sim_file — Function

get_sim_file(::Simulation, filename)

return the corresponding file of the given simulation "stdout" (or "-"), "stderr" and "" respectively retirect to stdout, stderr and devnull, other names are interpreted as file names