Systems and States

Systems

TensorMixedStates.System — Type

type System

represent a quantum system

Fields

sites::Vector{<:AbstractSite}: sites of the system

pure_indices::Vector{Index}: Indices for pure representations

mixed_indices::Vector{Index}: Indices for mixed representations

Examples

System(10, Qubit())
System([Qubit(), SpinOne(), Qubit(), Boson(5)])

Indexation

system[i]                  gives site i
SysIndex{Pure}(system, i)  gives pure index i
SysIndex{Mixed}(system, i) gives mixed index i

Base.length — Method

length(::System)

return the number of sites in the system

NDTensors.sim — Method

sim(::System)

create a clone of the system: identical but with different indices

TensorMixedStates.:⊗ — Method

::System ⊗ ::System
tensor(::System, ::System)

create the tensorial product of two systems

States

TensorMixedStates.Limits — Type

A type to hold MPS limits

Fields

cutoff: the cutoff under which singular values are neglected

maxdim: the maximum bond dimension

TensorMixedStates.State — Type

struct State{R <: PM}
State{R}(::System, states)
State{R}(::Int, ::AbstractSite, state)
State{R}(::Vector{<:AbstractSite}, state)
State(::State, ::MPS)

represent the complete state of the simulated quantum system

Type parameter

R is Pure or Mixed and represent the type of representation used

Fields

system::System: system description

state::MPS: system state

preobs::PreObs: preprocessing data for computing observables

Examples

State{Pure}(system, "Up")
State{Mixed}(system, ["Up", "Dn", "Up"])
State{Mixed}(system, "FullyMixed")
State{Pure}(system, [1, 0])
State{Pure}(10, Qubit(), "Up")
State{Mixed}([Qubit(), Boson(4), Fermion()], ["Up", "2", "Occ"])
State(state, mps)        returns a new state with the same system but a new mps

Operations

states can be added, substracted and multiplied by numbers

Base.length — Method

length(::State)

return the number of sites in the state

ITensorMPS.maxlinkdim — Method

maxlinkdim(::State)

return the maximum link dimension in the state

TensorMixedStates.mix — Function

mix(::Index)

return an ITensor.Index for a mixed representation corresponding to the pure representation Index given

mix(::State)

transform a pure representation into a mixed representation

Base.truncate — Method

truncate(::State; limits::Limits)

apply the truncations to the given state

TensorMixedStates.trace — Method

trace(::State)

Return the trace of the system, mostly usefull for mixed representations. This should be one.

TensorMixedStates.trace2 — Function

trace2(::State)

Return the trace of the square density matrix, mostly usefull for mixed representations. This is one for pure representations.

LinearAlgebra.norm — Method

norm(::State)

Return the norm of the state, mostly usefull for pure representations. This should be one for pure representation.

TensorMixedStates.hermitianize — Function

hermitianize(::State)

modify the state so that it is Hermitian (only useful for mixed state)

TensorMixedStates.hermitianity — Function

hermitianity(::State)

hermitianity measure whether density matrix for mixed state is Hermitian as it should.

return a value from 0 (anti Hermitian) to 1 (Hermitian)

return 1 for pure state

TensorMixedStates.RandomState — Type

RandomState{Pure|Mixed}(::System, linkdims::Int)
RandomState{Pure|Mixed}(::Int, ::AbstractSite, linkdims::Int)
RandomState{Pure|Mixed}(::Vector{<:AbstractSite}, linkdims::Int)
RandomState(::State, linkdims::Int)

Return a random state, with the specified link dimension. If a State is given, randomize the given state.

TensorMixedStates.partial_trace — Function

partial_trace(::State, ::Vector{Int} [; keepers = true])

return the state partially traced at the given positions, alternatively one can give the positions to keep by setting keepers = true

Simulations

TensorMixedStates.Simulation — Type

Simulation(state[, time = 0])
Simulation(sim, state[, time = sim.time])

A type to represent simulation data and store time and file data. It is used and returned by runTMS. The first form creates a simulation object. The second updates the state in the simulation object. (see also get_sim_file)

Most functions applicable to States can be applied to Simulations

Fields

state : the state of the system

time : the simulation time

output : if not nothing an io where to redirect output

files : a dictionary holding io or dict where to write data

data : a dictionary holding data collected for the Data objects

formats : format info for the output

TensorMixedStates.get_sim_file — Function

get_sim_file(::Simulation, filename)

return the corresponding file of the given simulation "stdout" (or "-"), "stderr" and "" respectively redirect to stdout, stderr and devnull, other names are interpreted as file names.

Filename finishing by ".json" will return a Dict where to store data and this data will be output in JSON format in the file by runTMS at the end.

Special filenames of the form Data(name) return a Dict where to store Data. Those Dict are gathered as a Dict in the data field of the Simulation